requirements.txt
requests
lxml
cssselect
argcomplete
progressbar2

